+++ embedded-graphics/CHANGELOG.md
@@ -15,6 +15,7 @@ Embedded Graphics is a `no_std` library for adding graphics features to display 
 ### Fixed
 
 - [#267](https://github.com/jamwaffles/embedded-graphics/pull/267) Fonts with CHARACTER_SPACING now properly render the background color between characters.
+- [#276](https://github.com/jamwaffles/embedded-graphics/pull/276) Fixed Chars like '°' and similar for `Font12x16` & `Font24x32`
 
 ## [0.6.0-beta.2] - 2020-03-06